Add setup_modifier_factory and state fixtures
