+++ tests/llmcompressor/pytorch/modifiers/pruning/wanda/test_pytorch.py
@@ -2,13 +2,11 @@ import pytest
 
 from llmcompressor.modifiers.factory import ModifierFactory
 from llmcompressor.modifiers.pruning.wanda import WandaPruningModifier
-from tests.llmcompressor.modifiers.conf import setup_modifier_factory
 
 
 @pytest.mark.unit
+@pytest.mark.usefixtures("setup_modifier_factory")
 def test_wanda_pytorch_is_registered():
-    setup_modifier_factory()
-
     sparsity = 0.5
     targets = "__ALL_PRUNABLE__"
