+++ tests/llmcompressor/pytorch/modifiers/smoothquant/test_pytorch.py
@@ -1,14 +1,7 @@
 import pytest
 from torch.nn import Linear
 
-from llmcompressor.core import State
 from llmcompressor.modifiers.smoothquant import SmoothQuantModifier
-from tests.llmcompressor.pytorch.helpers import LinearNet
-
-
-@pytest.fixture
-def state():
-    return State(model=LinearNet())
 
 
 @pytest.mark.unit
